Add files via upload
Here the python program will be connected to the excel spreadsheet which has names along with USNs of the students. When you run the program you can enter the date you want to take the attendance for giving multiple USNs as the input. By default rest of the USNs will be marked as absent.