Add files via upload
Here the python program will be connected to an excel spreadsheet which consists of names and USNs of students. While running the program we mention the date for which we are taking the attendance can give multiple USN as the input which will automatically update it to excel as 'Present' rest will be marked as 'Absent' by default.